client/backend/models.py
import logging
from datetime import timedelta
from typing import Any, List, Union
import uuid

from asgiref.sync import async_to_sync
from channels.layers import get_channel_layer
from django.apps import apps
from django.conf import settings
from django.db import models
from django.utils import timezone
